refactor: 调整 obsolecent hooks 路径
