+++ src/obsolescent/graph/indented.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import G6, { INode, IG6GraphEvent } from '@antv/g6';
 import ChartLoading from '../../util/createLoading';
 import { ErrorBoundary } from '../../base';
-import useGraph from '../../hooks/useGraph';
+import useGraph from '../hooks/useGraph';
 import { defaultNodeSize, defaultLabelCfg, defaultStateStyles } from './constants';
 import {
   processMinimap,
