+++ src/obsolescent/graph/organizationTreeGraph.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import G6, { NodeConfig } from '@antv/g6';
 import ChartLoading from '../../util/createLoading';
 import { ErrorBoundary } from '../../base';
-import useGraph from '../../hooks/useGraph';
+import useGraph from '../hooks/useGraph';
 import { customIconNode } from './customItems';
 import { defaultLabelCfg, defaultStateStyles } from './constants';
 import {
